pyproject.toml
[project]
name = "rss-parser"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "snusEbjoer", email = "mrkill2222@gmail.com" }
]
dependencies = [
    "bs4>=0.0.2",
    "beautifulsoup4>=4.12.3",
    "requests>=2.31.0",
    "feedparser>=6.0.11",
    "python-dateutil>=2.9.0.post0",
    "pytelegrambotapi>=4.17.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.5",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/rss_parser"]
